pyproject.toml
[project]
name = "wsproxy-test"
version = "0.1.0"
description = "Integration tests and load testing for wsproxy"
requires-python = ">=3.12"
dependencies = [
    "websockets>=12.0",
    "redis>=5.0.0",
    "aiohttp>=3.9.0",
    "fastapi>=0.109.0",
    "granian>=1.2.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "uvloop>=0.19.0",
    "orjson>=3.9.0",
    "pyjwt>=2.8.0",
]

[tool.uv]
dev-dependencies = []